refactor: reallocation of fixtures & rename clustering pipeline
