+++ tests/extraction/clustering/test_fdbscan_pipeline.py
@@ -1,8 +1,6 @@
 import pandas as pd
-import pytest
 from sklearn.pipeline import Pipeline
 
-from gps_activity.extraction.factory.clustering import FDBSCANFactory
 from gps_activity.models import DataFramePivotFields
 from gps_activity.models import DefaultValues
 
@@ -11,45 +9,7 @@ default_values = DefaultValues()
 pivot_fields = DataFramePivotFields()
 
 
-@pytest.fixture
-def clustering_gps_points():
-    x = list(range(10)) + [40, 80, 120] + list(range(150, 160))
-    y = list(range(len(x)))
-    unixtime = list(range(len(x)))
-    cluster_candidates = [True] * 5 + [False] * 5 + [True] * 13
-    df = pd.DataFrame(
-        {
-            pivot_fields.projected_lat: x,
-            pivot_fields.projected_lon: y,
-            pivot_fields.computed_unixtime: unixtime,
-            pivot_fields.fragmentation_output: cluster_candidates,
-        },
-    )
-    df["garbage1"] = "123"
-    df["garbage2"] = "123"
-    df[pivot_fields.source_vehicle_id] = "123"
-    return df
-
-
-@pytest.fixture
-def many_vehicle_gps(clustering_gps_points: pd.DataFrame):
-    df = clustering_gps_points
-    plate_no = df[pivot_fields.source_vehicle_id].to_list()
-    plate_no[-1] = "321"
-    df[pivot_fields.source_vehicle_id] = plate_no
-    return df
-
-
-@pytest.fixture
-def clustering_pipeline(gps_pivot_fields: DataFramePivotFields):
-    return FDBSCANFactory.factory_pipeline(
-        source_vehicle_id_column=gps_pivot_fields.source_vehicle_id,
-        eps=2,
-        min_samples=3,
-    )
-
-
-class TestPreprocessingPipelineFactory:
+class TestFDBSCAN:
 
     CLUSTER_OUPUT_COLUMNS = pivot_fields.clustering_output
     NOISE_IMPUTE_VALUE = default_values.noise_gps_cluster_id
@@ -66,15 +26,15 @@ class TestPreprocessingPipelineFactory:
         cluster_ids = gps.loc[~gps[fragmentation_flag], cluster_cand_col]
         assert (cluster_ids == self.NOISE_IMPUTE_VALUE).all()
 
-    def test_instance(self, clustering_pipeline: Pipeline):
-        assert isinstance(clustering_pipeline, Pipeline)
+    def test_instance(self, fdbscan_pipeline: Pipeline):
+        assert isinstance(fdbscan_pipeline, Pipeline)
 
     def test_pipeline(
         self,
         clustering_gps_points: pd.DataFrame,
-        clustering_pipeline: Pipeline,
+        fdbscan_pipeline: Pipeline,
     ):
-        y_pred = clustering_pipeline.fit_predict(clustering_gps_points)
+        y_pred = fdbscan_pipeline.fit_predict(clustering_gps_points)
         clustering_gps_points[self.CLUSTER_OUPUT_COLUMNS] = y_pred
         self.__test_cluster_count(
             clustering_gps_points,
@@ -89,10 +49,10 @@ class TestPreprocessingPipelineFactory:
     def test_unique_vehicle_constrain(
         self,
         many_vehicle_gps: pd.DataFrame,
-        clustering_pipeline: Pipeline,
+        fdbscan_pipeline: Pipeline,
     ):
         try:
-            clustering_pipeline.fit_predict(many_vehicle_gps)
+            fdbscan_pipeline.fit_predict(many_vehicle_gps)
             raise AssertionError("Must fail due to single vehicle constrain")
         except ValueError:
             assert True
